backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.20.0
sqlalchemy>=1.4.41
pymysql>=1.0.2
cryptography>=38.0.1
pydantic>=1.10.7
asyncio>=3.4.3
python-dotenv>=1.0.0
requests>=2.28.2
beautifulsoup4>=4.11.2
dashscope>=1.10.0
python-multipart>=0.0.5
pymilvus>=2.3.0
numpy>=1.24.0 